% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recebe_input.R
\name{recebe_input}
\alias{recebe_input}
\title{This function gets input from user}
\usage{
recebe_input()
}
\description{
This function gets input from user
}
\examples{
recebe_input()
}
\keyword{input}
\keyword{prompt,}
