% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_acento_arquivos.R
\name{remove_acento_arquivos}
\alias{remove_acento_arquivos}
\title{This function removes all accent of all files in the choosen folder.}
\usage{
remove_acento_arquivos(dir)
}
\description{
This function removes all accent of all files in the choosen folder.
}
\examples{
remove_acento_arquivos()
}
\keyword{acentos}
