% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_acento_arquivos.R
\name{remove_acento_arquivos}
\alias{remove_acento_arquivos}
\title{Removes all accent of all files in the choosen folder.}
\usage{
remove_acento_arquivos(dir)
}
\description{
Removes all accent of all files in the choosen folder.
}
\examples{
do not run!
my_folder <- getwd()
remove_acento_arquivos(my_folder)
}
\keyword{acentos}
