% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_acento_arquivos.R
\name{remove_acento_arquivos}
\alias{remove_acento_arquivos}
\title{uma função que remove acentos de todos os arquivos em um diretório}
\usage{
remove_acento_arquivos(dir)
}
\description{
uma função que remove acentos de todos os arquivos em um diretório
}
\examples{
remove_acento_arquivos()
}
\keyword{acentos}
